Try to resolve plugin sign failure
https://github.com/Gowee/traceroute-map-panel/runs/2146736283?check_suite_focus=true

Error:

> Run npx @grafana/toolkit plugin:sign
> Invalid command: plugin:sign
> See --help for a list of available commands.

The problem appears to be that npx attempts to install and then use another
outdated version of @grafana/toolkit in global somehow. The mysterious
behaviours can be reproduced on node:12 altough almost the same workflow
works well when releasing v0.3.0-beta2. Anyway, just wrap
@grafana/toolkit command in package.json commands. Besides, node version
for CI runner is also upgraded, to 15.